+++ backend/src/main/java/io/openems/backend/influx/InfluxdbQueryWrapper.java
@@ -1,4 +1,4 @@
-package io.openems.femsserver.influx;
+package io.openems.backend.influx;
 
 import java.time.Instant;
 import java.time.ZonedDateTime;
@@ -23,9 +23,9 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonNull;
 import com.google.gson.JsonObject;
 
-import io.openems.femsserver.utilities.Address;
-import io.openems.femsserver.utilities.JsonUtils;
-import io.openems.femsserver.utilities.OpenemsException;
+import io.openems.backend.utilities.Address;
+import io.openems.backend.utilities.JsonUtils;
+import io.openems.backend.utilities.OpenemsException;
 
 public class InfluxdbQueryWrapper {
