+++ backend/src/main/java/io/openems/backend/odoo/OdooModel.java
@@ -1,4 +1,4 @@
-package io.openems.femsserver.odoo;
+package io.openems.backend.odoo;
 
 import java.util.List;
