+++ backend/src/main/java/io/openems/backend/odoo/OdooObject.java
@@ -1,4 +1,4 @@
-package io.openems.femsserver.odoo;
+package io.openems.backend.odoo;
 
 import java.time.Instant;
 import java.time.LocalDateTime;
