+++ backend/src/main/java/io/openems/backend/utilities/Address.java
@@ -1,4 +1,4 @@
-package io.openems.femsserver.utilities;
+package io.openems.backend.utilities;
 
 public class Address {
 	private final String thingId;
