+++ backend/src/main/java/io/openems/backend/utilities/ManyToMany.java
@@ -1,4 +1,4 @@
-package io.openems.femsserver.utilities;
+package io.openems.backend.utilities;
 
 // Source: http://stackoverflow.com/questions/20390923/do-we-have-a-multibimap
 import java.util.Set;
