+++ backend/src/main/java/io/openems/backend/utilities/OpenemsException.java
@@ -1,4 +1,4 @@
-package io.openems.femsserver.utilities;
+package io.openems.backend.utilities;
 
 public class OpenemsException extends Exception {
 	private static final long serialVersionUID = 5015013132334439401L;
