+++ backend/src/main/java/io/openems/backend/utilities/StringUtils.java
@@ -1,4 +1,4 @@
-package io.openems.femsserver.utilities;
+package io.openems.backend.utilities;
 
 import com.google.gson.JsonObject;
