+++ backend/src/main/java/io/openems/backend/utilities/WebSocketUtils.java
@@ -1,4 +1,4 @@
-package io.openems.femsserver.utilities;
+package io.openems.backend.utilities;
 
 import org.java_websocket.WebSocket;
 import org.java_websocket.exceptions.WebsocketNotConnectedException;
